Super early concept of logwatcher based log parsing
Hardcoded testing nonsense everywhere and logs tailored for 2 different
sites that have 2 different custom log formats.

I abandoned this idea when it turned out to take a lot of CPU (though in
retrospect I was testing a non-optimized debug compile, so it may not be
fair), and I hate the idea of `tail -f` parsing in general.