Third attempt: Don't accumulate stats ourselves, instead pass it off to statsd
This still uses the log parsing from the 2nd attempt, but by letting statsd
do the accumulating I can remove all the threading and locking stuff, along
with the stats accumulation struct. This makes it much more straightforward,
and lets it be compatible with anything that implements the statsd protocol.
I'm using Telegraf, myself.